+++ src/main/java/me/superischroma/spectaculation/item/armor/hardened/HardenedDiamondLeggings.java
@@ -29,7 +29,7 @@ public class HardenedDiamondLeggings implements ToolStatistics, MaterialFunction
     }
     
     @Override
-    public int getBaseDefense()
+    public double getBaseDefense()
     {
         return 95;
     }
