+++ src/main/java/me/superischroma/spectaculation/item/armor/vanilla/diamond/DiamondLeggings.java
@@ -29,7 +29,7 @@ public class DiamondLeggings implements ToolStatistics, MaterialFunction
     }
 
     @Override
-    public int getBaseDefense()
+    public double getBaseDefense()
     {
         return 30;
     }
