+++ src/main/java/me/superischroma/spectaculation/item/armor/vanilla/leather/LeatherChestplate.java
@@ -29,7 +29,7 @@ public class LeatherChestplate implements ToolStatistics, MaterialFunction
     }
 
     @Override
-    public int getBaseDefense()
+    public double getBaseDefense()
     {
         return 15;
     }
