+++ src/main/java/me/superischroma/spectaculation/item/armor/vanilla/leather/LeatherLeggings.java
@@ -29,7 +29,7 @@ public class LeatherLeggings implements ToolStatistics, MaterialFunction
     }
 
     @Override
-    public int getBaseDefense()
+    public double getBaseDefense()
     {
         return 10;
     }
