+++ src/main/java/me/superischroma/spectaculation/item/dragon/strong/StrongDragonBoots.java
@@ -9,19 +9,19 @@ import me.superischroma.spectaculation.item.armor.LeatherArmorStatistics;
 public class StrongDragonBoots implements MaterialFunction, LeatherArmorStatistics
 {
     @Override
-    public int getBaseStrength()
+    public double getBaseStrength()
     {
         return 25;
     }
 
     @Override
-    public int getBaseHealth()
+    public double getBaseHealth()
     {
         return 60;
     }
 
     @Override
-    public int getBaseDefense()
+    public double getBaseDefense()
     {
         return 90;
     }
