+++ src/main/java/me/superischroma/spectaculation/item/dragon/strong/StrongDragonChestplate.java
@@ -9,19 +9,19 @@ import me.superischroma.spectaculation.item.armor.LeatherArmorStatistics;
 public class StrongDragonChestplate implements MaterialFunction, LeatherArmorStatistics
 {
     @Override
-    public int getBaseStrength()
+    public double getBaseStrength()
     {
         return 25;
     }
 
     @Override
-    public int getBaseHealth()
+    public double getBaseHealth()
     {
         return 120;
     }
 
     @Override
-    public int getBaseDefense()
+    public double getBaseDefense()
     {
         return 160;
     }
