+++ src/main/java/me/superischroma/spectaculation/item/dragon/superior/SuperiorDragonBoots.java
@@ -9,7 +9,7 @@ import me.superischroma.spectaculation.item.armor.LeatherArmorStatistics;
 public class SuperiorDragonBoots implements MaterialFunction, LeatherArmorStatistics
 {
     @Override
-    public int getBaseStrength()
+    public double getBaseStrength()
     {
         return 10;
     }
@@ -27,7 +27,7 @@ public class SuperiorDragonBoots implements MaterialFunction, LeatherArmorStatis
     }
 
     @Override
-    public int getBaseIntelligence()
+    public double getBaseIntelligence()
     {
         return 25;
     }
@@ -39,13 +39,13 @@ public class SuperiorDragonBoots implements MaterialFunction, LeatherArmorStatis
     }
 
     @Override
-    public int getBaseHealth()
+    public double getBaseHealth()
     {
         return 80;
     }
 
     @Override
-    public int getBaseDefense()
+    public double getBaseDefense()
     {
         return 110;
     }
