+++ src/main/java/me/superischroma/spectaculation/item/dragon/young/YoungDragonChestplate.java
@@ -15,13 +15,13 @@ public class YoungDragonChestplate implements MaterialFunction, LeatherArmorStat
     }
 
     @Override
-    public int getBaseHealth()
+    public double getBaseHealth()
     {
         return 120;
     }
 
     @Override
-    public int getBaseDefense()
+    public double getBaseDefense()
     {
         return 160;
     }
