+++ src/main/java/me/superischroma/spectaculation/item/dragon/young/YoungDragonHelmet.java
@@ -11,13 +11,13 @@ public class YoungDragonHelmet implements MaterialFunction, SkullStatistics, Too
     }
 
     @Override
-    public int getBaseHealth()
+    public double getBaseHealth()
     {
         return 70;
     }
 
     @Override
-    public int getBaseDefense()
+    public double getBaseDefense()
     {
         return 110;
     }
