+++ src/main/java/me/superischroma/spectaculation/item/orb/PlasmafluxPowerOrb.java
@@ -77,7 +77,7 @@ public class PlasmafluxPowerOrb extends PowerOrb
                 return null;
             }
             @Override
-            public int getBaseStrength()
+            public double getBaseStrength()
             {
                 return 35;
             }
