+++ src/main/java/me/superischroma/spectaculation/item/weapon/LeapingSword.java
@@ -17,7 +17,7 @@ public class LeapingSword implements ToolStatistics, MaterialFunction, Ability
     }
 
     @Override
-    public int getBaseStrength()
+    public double getBaseStrength()
     {
         return 100;
     }
