+++ src/main/java/me/superischroma/spectaculation/reforge/RapidReforge.java
@@ -15,9 +15,9 @@ public class RapidReforge implements Reforge
     }
 
     @Override
-    public RarityValue<Integer> getStrength()
+    public RarityValue<Double> getStrength()
     {
-        return new RarityValue<>(2, 3, 4, 7, 10, 15);
+        return new RarityValue<>(2.0, 3.0, 4.0, 7.0, 10.0, 15.0);
     }
 
     @Override
